rally_backup_2025_06_05_1439/rally_backup_2025_06_04_1630/requirements.txt
flask==3.0.2
flask-socketio==5.3.6
pandas==2.2.1
flask-cors==4.0.0
python-dotenv==1.0.1
openai>=1.2.3
selenium==4.18.1
urllib3<2.0.0
beautifulsoup4==4.12.3
webdriver-manager==4.0.1
requests==2.31.0
Flask-Assets==2.1.0
libsass==0.22.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
alembic>=1.13.1
